requirements.txt
# Core Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.12.0

# Environment & Configuration
python-dotenv>=1.0.0

# Background Tasks & Queue
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0

# AI/ML Dependencies (updated for Python 3.10+ compatibility)
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
openai>=1.12.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0

# Additional Utilities
aiofiles>=23.2.0
structlog>=23.2.0
pydantic-settings>=2.0.0

# Monitoring & Metrics
prometheus-client>=0.19.0
psutil>=5.9.0

# OpenTelemetry
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-celery>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
opentelemetry-exporter-prometheus>=0.42b0